refactor: simplify PXI credential form rendering
