+++ app/src/components/generative/ProviderServerCredentialsPanel.tsx
@@ -6,149 +6,36 @@ import { graphql, useLazyLoadQuery, useMutation } from "react-relay";
 import {
   Alert,
   Button,
-  CredentialField,
-  CredentialInput,
-  Dialog,
-  DialogCloseButton,
-  DialogContent,
-  DialogHeader,
-  DialogTitle,
-  DialogTitleExtra,
   Flex,
   Form,
-  Label,
   RedactedCredentialField,
   Text,
-  ToggleButton,
-  ToggleButtonGroup,
   View,
 } from "@phoenix/components";
 import { useNotifySuccess, useViewer } from "@phoenix/contexts";
-import { useCredentialsContext } from "@phoenix/contexts/CredentialsContext";
-import { isModelProvider } from "@phoenix/utils/generativeUtils";
 
-import type { ProviderCredentialsDialogSecretsQuery } from "./__generated__/ProviderCredentialsDialogSecretsQuery.graphql";
-import type { ProviderCredentialsDialogUpsertOrDeleteSecretsMutation } from "./__generated__/ProviderCredentialsDialogUpsertOrDeleteSecretsMutation.graphql";
-
-type CredentialViewType = "browser" | "secrets";
+import type { ProviderServerCredentialsPanelSecretsQuery } from "./__generated__/ProviderServerCredentialsPanelSecretsQuery.graphql";
+import type { ProviderServerCredentialsPanelUpsertOrDeleteSecretsMutation } from "./__generated__/ProviderServerCredentialsPanelUpsertOrDeleteSecretsMutation.graphql";
 
 type ProviderCredentialRequirement = {
   readonly envVarName: string;
   readonly isRequired: boolean;
 };
 
-export type ProviderCredentialsDialogProvider = {
+export type ProviderServerCredentialsPanelProvider = {
   readonly name: string;
   readonly key: string;
   readonly credentialRequirements: readonly ProviderCredentialRequirement[];
 };
 
-type ProviderCredentialsDialogMode = "browser-and-server" | "server-only";
-
 // Form values type for react-hook-form
 type ServerCredentialsFormValues = Record<string, string>;
-
-export function ProviderCredentialsDialog({
-  provider,
-  mode = "browser-and-server",
-  onCredentialsUpdated,
-}: {
-  provider: ProviderCredentialsDialogProvider;
-  mode?: ProviderCredentialsDialogMode;
-  onCredentialsUpdated?: () => void;
-}) {
-  const { viewer } = useViewer();
-  const isAdmin = !viewer || viewer.role?.name === "ADMIN";
-  const [credentialView, setCredentialView] =
-    useState<CredentialViewType>("browser");
-  const isServerOnly = mode === "server-only";
-
-  return (
-    <Dialog>
-      <DialogContent>
-        <DialogHeader>
-          <DialogTitle>
-            Configure {isAdmin || isServerOnly ? "" : "Local "}
-            {provider.name} Credentials
-          </DialogTitle>
-          <DialogTitleExtra>
-            <DialogCloseButton slot="close" />
-          </DialogTitleExtra>
-        </DialogHeader>
-        <View padding="size-200">
-          {isServerOnly ? (
-            <ProviderServerCredentialsPanel
-              provider={provider}
-              onCredentialsUpdated={onCredentialsUpdated}
-            />
-          ) : isAdmin ? (
-            <Flex direction="column" gap="size-200">
-              <ToggleButtonGroup
-                selectedKeys={[credentialView]}
-                size="S"
-                aria-label="Credential Source"
-                onSelectionChange={(v) => {
-                  if (v.size === 0) {
-                    return;
-                  }
-                  const view = v.keys().next().value as CredentialViewType;
-                  if (view === "browser" || view === "secrets") {
-                    setCredentialView(view);
-                  }
-                }}
-              >
-                <ToggleButton aria-label="Browser" id="browser">
-                  Browser
-                </ToggleButton>
-                <ToggleButton aria-label="Secrets" id="secrets">
-                  Secrets
-                </ToggleButton>
-              </ToggleButtonGroup>
-              {credentialView === "browser" && (
-                <>
-                  <View paddingBottom="size-100">
-                    <Text size="XS" color="text-700">
-                      Credentials stored in your browser. Only you can see
-                      these, and they are sent with each API request.
-                    </Text>
-                  </View>
-                  <Form>
-                    <BrowserCredentials provider={provider} />
-                  </Form>
-                </>
-              )}
-              {credentialView === "secrets" && (
-                <ServerCredentialsSection
-                  provider={provider}
-                  onCredentialsUpdated={onCredentialsUpdated}
-                />
-              )}
-            </Flex>
-          ) : (
-            <>
-              <View paddingBottom="size-100">
-                <Text size="XS">
-                  Set the credentials for the {provider.name} API. These
-                  credentials will be stored entirely in your browser and will
-                  only be sent to the server during API requests.
-                </Text>
-              </View>
-              <Form>
-                <BrowserCredentials provider={provider} />
-              </Form>
-            </>
-          )}
-        </View>
-      </DialogContent>
-    </Dialog>
-  );
-}
 
 export function ProviderServerCredentialsPanel({
   provider,
   onCredentialsUpdated,
 }: {
-  provider: ProviderCredentialsDialogProvider;
+  provider: ProviderServerCredentialsPanelProvider;
   onCredentialsUpdated?: () => void;
 }) {
   const { viewer } = useViewer();
@@ -175,7 +62,7 @@ function ServerCredentialsSection({
   provider,
   onCredentialsUpdated,
 }: {
-  provider: ProviderCredentialsDialogProvider;
+  provider: ProviderServerCredentialsPanelProvider;
   onCredentialsUpdated?: () => void;
 }) {
   return (
@@ -198,76 +85,11 @@ function ServerCredentialsSection({
   );
 }
 
-function BrowserCredentials({
-  provider,
-}: {
-  provider: ProviderCredentialsDialogProvider;
-}) {
-  const providerKey = provider.key;
-  const isValidProvider = isModelProvider(providerKey);
-
-  const { setCredential, credentials } = useCredentialsContext((state) => ({
-    setCredential: state.setCredential,
-    credentials: isValidProvider ? state[providerKey] : undefined,
-  }));
-  const credentialRequirements = provider.credentialRequirements;
-
-  const clearLocalCredentials = () => {
-    if (!isValidProvider) return;
-    provider.credentialRequirements.forEach(({ envVarName }) => {
-      setCredential({
-        provider: providerKey,
-        envVarName,
-        value: "",
-      });
-    });
-  };
-
-  if (!isValidProvider) {
-    return <Text color="warning">Unknown provider type: {providerKey}</Text>;
-  }
-
-  if (provider.credentialRequirements.length === 0) {
-    return <Text color="text-700">Browser credentials are not required.</Text>;
-  }
-
-  return (
-    <Flex direction="column" gap="size-100">
-      {credentialRequirements.map(({ envVarName, isRequired }) => (
-        <CredentialField
-          key={envVarName}
-          isRequired={isRequired}
-          onChange={(value) => {
-            setCredential({
-              provider: providerKey,
-              envVarName,
-              value,
-            });
-          }}
-          value={credentials?.[envVarName] ?? ""}
-        >
-          <Label>{envVarName}</Label>
-          <CredentialInput />
-        </CredentialField>
-      ))}
-      <Button
-        onPress={clearLocalCredentials}
-        css={css`
-          align-self: flex-start;
-          margin-top: var(--global-dimension-size-100);
-        `}
-      >
-        Clear Local Credentials
-      </Button>
-    </Flex>
-  );
-}
-
 function ServerCredentials({
   provider,
   onCredentialsUpdated,
 }: {
-  provider: ProviderCredentialsDialogProvider;
+  provider: ProviderServerCredentialsPanelProvider;
   onCredentialsUpdated?: () => void;
 }) {
   const notifySuccess = useNotifySuccess();
@@ -275,30 +97,31 @@ function ServerCredentials({
   const [fetchKey, setFetchKey] = useState(0);
   const secretKeys = provider.credentialRequirements.map((c) => c.envVarName);
 
-  const secretsData = useLazyLoadQuery<ProviderCredentialsDialogSecretsQuery>(
-    graphql`
-      query ProviderCredentialsDialogSecretsQuery($secretKeys: [String!]!) {
-        secrets(keys: $secretKeys) {
-          edges {
-            node {
-              key
-              value {
-                __typename
-                ... on DecryptedSecret {
-                  value
-                }
-                ... on UnparsableSecret {
-                  parseError
+  const secretsData =
+    useLazyLoadQuery<ProviderServerCredentialsPanelSecretsQuery>(
+      graphql`
+        query ProviderServerCredentialsPanelSecretsQuery($secretKeys: [String!]!) {
+          secrets(keys: $secretKeys) {
+            edges {
+              node {
+                key
+                value {
+                  __typename
+                  ... on DecryptedSecret {
+                    value
+                  }
+                  ... on UnparsableSecret {
+                    parseError
+                  }
                 }
               }
             }
           }
         }
-      }
-    `,
-    { secretKeys },
-    { fetchKey, fetchPolicy: "store-and-network" }
-  );
+      `,
+      { secretKeys },
+      { fetchKey, fetchPolicy: "store-and-network" }
+    );
 
   const serverSecretMap = new Map<string, string>();
   const unparsableSecrets = new Map<string, string>();
@@ -339,9 +162,9 @@ function ServerCredentials({
   );
 
   const [commit, isCommitting] =
-    useMutation<ProviderCredentialsDialogUpsertOrDeleteSecretsMutation>(
+    useMutation<ProviderServerCredentialsPanelUpsertOrDeleteSecretsMutation>(
       graphql`
-        mutation ProviderCredentialsDialogUpsertOrDeleteSecretsMutation(
+        mutation ProviderServerCredentialsPanelUpsertOrDeleteSecretsMutation(
           $input: UpsertOrDeleteSecretsMutationInput!
         ) {
           upsertOrDeleteSecrets(input: $input) {
